+++ src/main/java/io/lettuce/core/api/sync/RedisSortedSetCommands.java
@@ -279,22 +279,26 @@ public interface RedisSortedSetCommands<K, V> {
     List<V> zinter(ZAggregateArgs aggregateArgs, K... keys);
 
     /**
-     * This command is similar to ZINTER, but instead of returning the result set, it returns just the cardinality of the result.
+     * This command is similar to {@link #zinter(java.lang.Object[])}, but instead of returning the result set, it returns just
+     * the cardinality of the result.
      *
      * @param keys the keys.
      * @return Long Integer reply the number of elements in the resulting intersection.
+     * @since 6.2
      */
     Long zintercard(K... keys);
 
     /**
-     * This command is similar to ZINTER, but instead of returning the result set, it returns just the cardinality of the result.
+     * This command is similar to {@link #zinter(java.lang.Object[])}, but instead of returning the result set, it returns just
+     * the cardinality of the result.
      *
      * @param limit If the intersection cardinality reaches limit partway through the computation, the algorithm will exit and
      *        yield limit as the cardinality
      * @param keys the keys.
      * @return Long Integer reply the number of elements in the resulting intersection.
+     * @since 6.2
      */
-    Long zintercard(int limit, K... keys);
+    Long zintercard(long limit, K... keys);
 
     /**
      * Intersect multiple sorted sets and returns the resulting sorted.
